PRSD-1354: refactor
